fullnode sdk address auto detect
