+++ scripts/sdkcom/__init__.py
@@ -1,4 +1,6 @@
+from config import *
 from common import *
-from rpc_call import *
-from tx_load import *
 from constant import *
+from rpc_call import *
+from actions import *
+from tx_load import *
